server/requirements.txt
# ── FastAPI & Server ──────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20       # Required for file uploads

# ── Database ─────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0                # Async PostgreSQL driver
alembic==1.14.1                # Database migrations

# ── Validation & Settings ────────────────────────────
pydantic[email]==2.10.4
pydantic-settings==2.7.1       # Load settings from .env

# ── Authentication ─────────────────────────────────
python-jose[cryptography]==3.5.0  # JWT token creation/verification
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt==4.1.3                     # Bcrypt backend (compatible with passlib)

# ── AI / LLM Providers ──────────────────────────────
google-generativeai>=0.8.0     # Google Gemini API (supports 2.5 Flash)
groq>=0.15.0                   # Groq API (Llama 3.3)
httpx==0.28.1                  # HTTP client (for Ollama + general use)

# ── Document Parsing ────────────────────────────────
PyMuPDF==1.25.3                # PDF text extraction
python-docx==1.1.2             # DOCX text extraction

# ── Export ───────────────────────────────────────────
openpyxl==3.1.5                # Excel export support

# ── Testing ──────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0         # Async test support
httpx==0.28.1                  # Async test client

# ── Dev Tools ────────────────────────────────────────
ruff==0.8.6                    # Fast Python linter & formatter
